update:order core service

+++ order-services/order-provider/src/main/java/com/mall/order/dal/entitys/OrderItem.java
@@ -19,85 +19,46 @@ import javax.persistence.*;
 public class OrderItem {
 
     @Id
-
     private String id;
 
-
-
     /**
-
      * 商品id
-
      */
-
     @Column(name = "item_id")
 
     private Long itemId;
 
-
-
     /**
-
      * 订单id
-
      */
-
     @Column(name = "order_id")
-
     private String orderId;
 
-
-
     /**
-
      * 商品购买数量
-
      */
-
     private Integer num;
 
-
-
     /**
-
      * 商品标题
-
      */
-
     private String title;
 
-
-
     /**
-
      * 商品单价
-
      */
-
-    private Double price;
-
-
+    private BigDecimal price;
 
     /**
-
      * 商品总金额
-
      */
-
     @Column(name = "total_fee")
-
-    private Double totalFee;
-
-
+    private BigDecimal totalFee;
 
     /**
-
      * 商品图片地址
-
      */
-
     @Column(name = "pic_path")
-
     private String picPath;
 
     /**
